separated CRUD handlers, completed SAM template
