+++ README.md
@@ -3,8 +3,6 @@
 ## Transaction Logging Service
 
 ## Pre-requisites
-- [Maven](https://maven.apache.org/) 
-- [OpenJDK 8](https://developers.redhat.com/products/openjdk/download)
 - [AWS SAM](https://docs.aws.amazon.com/serverless-application-model/latest/developerguide/serverless-sam-cli-install.html)
 - [Docker](https://www.docker.com/products/docker-desktop)
 
@@ -24,8 +22,12 @@ Starts the REST API on `http://127.0.0.1:3000`
 
 
 ### Available Endpoints
-
+POST, PUT
 `/transaction`
+
+GET, DELETE
+`/transaction/{UUID}/{Timestamp}`
+
 
 ## Example POST Request
 ```
@@ -33,17 +35,13 @@ http://127.0.0.1:3000/transaction?TableName=TransactionLogs
 
 {
     "TableName": "TransactionLogs",
-        "Item": {
-            "UUID": "110",
-            "Timestamp": 1605873564,
-            "Customer": "test man",
-            "ServiceType": "Proofing Services",
-            "CSP": "Idemia",
-            "Result": "in-progress",
-            "ProofingStatus": "attending in person proofing"
-
-        }
-    
+    "ServiceType": "Proofing Services",
+    "CSP": "Idemia",
+    "Result": "in-progress",
+    "ProofingStatus": "attending in person proofing"
 }
 
 ```
+TableName is the only required attribute, currently considering separating customers into different tables.
+
+UUID is derived from the `context.awsRequestId;`, but for testing the PUT Request endpoint you can comment out the UUID and Timestamp.
