--- transactionlogs/package.json
@@ -1,9 +1,0 @@
-{
-  "name": "nodejs-upgrade-functions",
-  "version": "1.0.0",
-  "private": true,
-  "dependencies": {
-    "dynamodb-doc": "^1.0.0",
-    "throat": "^4.0.0"
-  }
-}
